pyproject.toml
[tool.poetry]
name = "interfaceautotest"
version = "0.1.0"
description = ""
authors = ["NianJiu <t6i888@163.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"


requests = "^2.32.3"
pyyaml = "^6.0.2"
pytest = "^8.3.3"
jsonpath = "^0.82.2"
pymysql = "^1.1.1"
pytest-result-log = "^1.2.2"
allure-pytest = "^2.13.5"
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[[tool.poetry.source]]
name = "tsinghua"
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
priority = "primary"